PublicRoute has been used to protect the login and register routes. The ProtectedRoute component has been updated to redirect to the selection dashboard page if the user is authenticated. The App component has been updated to use the PublicRoute and ProtectedRoute components. The routes.jsx file has been updated to use the PublicRoute and ProtectedRoute components.
